--- android/src/main/java/org/lovebing/reactnative/baidumap/BaiduGuidePackage.java
@@ -1,41 +1,0 @@
-package org.lovebing.reactnative.baidumap;
-
-import android.content.Context;
-
-import com.facebook.react.ReactPackage;
-import com.facebook.react.bridge.JavaScriptModule;
-import com.facebook.react.bridge.NativeModule;
-import com.facebook.react.bridge.ReactApplicationContext;
-import com.facebook.react.uimanager.ViewManager;
-
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-
-public class BaiduGuidePackage implements ReactPackage {
-
-    private Context mContext;
-
-    public BaiduGuidePackage(Context context) {
-        this.mContext = context;
-    }
-
-    @Override
-    public List<NativeModule> createNativeModules(ReactApplicationContext reactContext) {
-        return Arrays.<NativeModule>asList(
-            new GuideModule(reactContext)
-        );
-    }
-
-    @Override
-    public List<ViewManager> createViewManagers(
-            ReactApplicationContext reactContext) {
-        return Arrays.<ViewManager>asList(
-
-        );
-    }
-
-    public List<Class<? extends JavaScriptModule>> createJSModules() {
-        return Collections.emptyList();
-    }
-}
