+++ android/src/main/java/org/lovebing/reactnative/baidumap/GeolocationModule.java
@@ -1,6 +1,7 @@
 package org.lovebing.reactnative.baidumap;
 
 import android.util.Log;
+import android.widget.Toast;
 
 import com.baidu.location.BDLocation;
 import com.baidu.location.BDLocationListener;
@@ -14,6 +15,16 @@ import com.baidu.mapapi.search.geocode.GeoCoder;
 import com.baidu.mapapi.search.geocode.OnGetGeoCoderResultListener;
 import com.baidu.mapapi.search.geocode.ReverseGeoCodeOption;
 import com.baidu.mapapi.search.geocode.ReverseGeoCodeResult;
+import com.baidu.mapapi.search.route.BikingRouteResult;
+import com.baidu.mapapi.search.route.DrivingRoutePlanOption;
+import com.baidu.mapapi.search.route.DrivingRouteResult;
+import com.baidu.mapapi.search.route.IndoorRouteResult;
+import com.baidu.mapapi.search.route.MassTransitRouteResult;
+import com.baidu.mapapi.search.route.OnGetRoutePlanResultListener;
+import com.baidu.mapapi.search.route.PlanNode;
+import com.baidu.mapapi.search.route.RoutePlanSearch;
+import com.baidu.mapapi.search.route.TransitRouteResult;
+import com.baidu.mapapi.search.route.WalkingRouteResult;
 import com.baidu.mapapi.utils.CoordinateConverter;
 import com.facebook.react.bridge.Arguments;
 import com.facebook.react.bridge.ReactApplicationContext;
@@ -25,6 +36,8 @@ public class GeolocationModule extends BaseModule
 
     private LocationClient locationClient;
     private static GeoCoder geoCoder;
+    RoutePlanSearch mSearch;
+    ReactApplicationContext context;
 
     public GeolocationModule(ReactApplicationContext reactContext) {
         super(reactContext);
@@ -48,10 +61,7 @@ public class GeolocationModule extends BaseModule
         Log.i("locationClient", "locationClient");
         locationClient.registerLocationListener(this);
     }
-    /**
-     *
-     * @return
-     */
+
     protected GeoCoder getGeoCoder() {
         if(geoCoder != null) {
             geoCoder.destroy();
@@ -61,11 +71,7 @@ public class GeolocationModule extends BaseModule
         return geoCoder;
     }
 
-    /**
-     *
-     * @param sourceLatLng
-     * @return
-     */
+
     protected LatLng getBaiduCoorFromGPSCoor(LatLng sourceLatLng) {
         CoordinateConverter converter = new CoordinateConverter();
         converter.from(CoordinateConverter.CoordType.GPS);
@@ -80,14 +86,78 @@ public class GeolocationModule extends BaseModule
         if(locationClient == null) {
             initLocationClient();
         }
-        Log.i("getCurrentPosition", "getCurrentPosition");
         locationClient.start();
     }
+
+
+    @ReactMethod
+    public void getGuideStart(int type) {
+        if(mSearch == null) {
+            initGuideType(type);
+        }else{
+            initGuideType(type);
+        }
+    }
+
+
+    private void initGuideType(int type) {
+        mSearch = RoutePlanSearch.newInstance();
+        guideType(type);
+    }
+
+
+
+    public void guideType(int type){
+        Toast.makeText(context,"hell0"+type,Toast.LENGTH_LONG).show();
+        PlanNode stNode = PlanNode.withCityNameAndPlaceName("杭州", "福鼎家园");
+        PlanNode enNode = PlanNode.withCityNameAndPlaceName("杭州", "西城广场");
+        mSearch.setOnGetRoutePlanResultListener(new OnGetRoutePlanResultListener() {
+
+            @Override
+            public void onGetWalkingRouteResult(WalkingRouteResult walkingRouteResult) {
+
+            }
+
+            @Override
+            public void onGetTransitRouteResult(TransitRouteResult transitRouteResult) {
+
+            }
+
+            @Override
+            public void onGetMassTransitRouteResult(MassTransitRouteResult massTransitRouteResult) {
+
+            }
+
+            @Override
+            public void onGetDrivingRouteResult(DrivingRouteResult result) {
+                Toast.makeText(context,"hell0",Toast.LENGTH_LONG).show();
+            }
+
+            @Override
+            public void onGetIndoorRouteResult(IndoorRouteResult indoorRouteResult) {
+
+            }
+
+            @Override
+            public void onGetBikingRouteResult(BikingRouteResult bikingRouteResult) {
+
+            }
+
+        });
+        mSearch.drivingSearch((new DrivingRoutePlanOption()).from(stNode).to(enNode));
+
+    }
+
+
     @ReactMethod
     public void geocode(String city, String addr) {
         getGeoCoder().geocode(new GeoCodeOption()
                 .city(city).address(addr));
     }
+
+
+
+
 
     @ReactMethod
     public void reverseGeoCode(double lat, double lng) {
@@ -155,4 +225,12 @@ public class GeolocationModule extends BaseModule
         }
         sendEvent("onGetReverseGeoCodeResult", params);
     }
+
+
+
+
+
+
+
+
 }
