+++ android/src/main/java/org/lovebing/reactnative/baidumap/User.java
@@ -1,0 +1,11 @@
+package org.lovebing.reactnative.baidumap;
+
+public class User {
+    private String name;
+    private String phone;
+
+    public User(String name,String phone){
+        this.name=name;
+        this.phone=phone;
+    }
+}
